pyproject.toml
[tool.poetry]
name = "telegramchatbotai"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.4.1"
beautifulsoup4 = "^4.12.3"
pydantic-settings = "^2.2.1"
pandas = "2.1.4"
scikit-learn = "^1.4.1.post1"
tqdm = "^4.66.2"
nltk = "^3.8.1"
openpyxl = "^3.1.2"
pyyaml = "^6.0.1"
pyspellchecker = "^0.8.1"
deep-translator = "^1.11.4"
lxml = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
